requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
langchain-groq
langgraph==0.2.28
langchain-core==0.3.10
chromadb 
datasets==3.0.1
pandas==2.2.3
numpy
tqdm==4.66.5
pydantic==2.9.2
httpx==0.27.2

# Optional: Observability with Arize Phoenix
# arize-phoenix[otel]==4.5.0
# opentelemetry-sdk==1.27.0
